pipeline/requirements.txt
anthropic>=0.43.0
python-docx>=1.1.0
pydub>=0.25.1
mutagen>=1.47.0
Pillow>=10.0.0
pyloudnorm>=0.1.1
soundfile>=0.12.1
pyyaml>=6.0
rich>=13.0
click>=8.1.0
jinja2>=3.1.0
feedgen>=1.0.0
